Moved FrameworkNode to inference_engine_transformations library
