+++ inference-engine/src/readers/ir_reader/CMakeLists.txt
@@ -33,6 +33,7 @@ target_link_libraries(${TARGET_NAME} PRIVATE ${NGRAPH_LIBRARIES}
                                              inference_engine_reader_api
                                              inference_engine_plugin_api
                                              inference_engine
+                                             inference_engine_transformations
                                              pugixml
                                              openvino::itt)
